+++ src/main/java/com/wootube/ioi/service/support/Config.java
@@ -1,4 +1,4 @@
-package com.wootube.ioi.web.controller.support;
+package com.wootube.ioi.service.support;
 
 import org.modelmapper.ModelMapper;
 
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class config {
+public class Config {
 	@Bean
 	public ModelMapper modelMapper() {
 		return new ModelMapper();
