--- src/main/java/com/wootube/ioi/web/controller/TestController.java
@@ -1,21 +1,0 @@
-package com.wootube.ioi.web.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@RequestMapping("/test")
-@Controller
-public class TestController {
-
-    @GetMapping
-    public String test() {
-        return "index.html";
-    }
-
-    @GetMapping("/video")
-    public String video() {
-        return "video.html";
-    }
-
-}
