+++ src/main/java/com/wootube/ioi/web/controller/VideoController.java
@@ -1,6 +1,6 @@
 package com.wootube.ioi.web.controller;
 
-import com.wootube.ioi.request.VideoDto;
+import com.wootube.ioi.service.dto.VideoRequestDto;
 import com.wootube.ioi.service.VideoService;
 
 import org.springframework.stereotype.Controller;
@@ -23,9 +23,9 @@ public class VideoController {
 	}
 
 	@PostMapping("/new")
-	public String video(MultipartFile uploadFile, VideoDto videoDto) {
-		videoService.create(uploadFile, videoDto);
-		return "redirect:/";
+	public String video(MultipartFile uploadFile, VideoRequestDto videoRequestDto) {
+		videoRequestDto = videoService.create(uploadFile, videoRequestDto);
+		return "redirect:/videos/" + videoRequestDto.getId();
 	}
 
 	@GetMapping("/{id}")
@@ -41,8 +41,8 @@ public class VideoController {
 	}
 
 	@PutMapping("/{id}")
-	public String updateVideo(@PathVariable Long id, MultipartFile uploadFile, VideoDto videoDto) {
-		videoService.update(id, uploadFile, videoDto);
+	public String updateVideo(@PathVariable Long id, MultipartFile uploadFile, VideoRequestDto videoRequestDto) {
+		videoService.update(id, uploadFile, videoRequestDto);
 		return "redirect:/videos/"+id;
 	}
 }
